# api/api-doc.yml
openapi: 3.0.3
info:
  title: Cloudflare 邮箱管理系统 API
  version: 3.0.0
  description: |
    Cloudflare Email Manager (CEM) 的统一 API 规范。系统已精简为 **单管理员模式**，不再开放多用户注册与邮箱申请，所有管理操作均由管理员完成。

    - 所有接口前缀统一为 `/api`
    - 统一使用 HttpOnly Cookie 维持会话
    - localStorage 仅缓存 UI 状态，业务数据一律来自接口
    - `allow_registration` 字段在健康检查与系统配置中继续返回，用于未来扩展（当前永远为只读）
servers:
  - url: http://localhost:8787
    description: 开发环境
  - url: https://cem.xrsec.workers.dev
    description: 生产环境
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
...
  /api/emails:
    get:
      tags: [Emails]
      summary: 获取邮件列表
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema: { type: string }
          description: 主题 / 发件人模糊搜索
        - name: status
          in: query
          schema: { type: string, enum: ['unread','read'] }
        - name: scope
          in: query
          schema: { type: string, enum: ['all'] }
          description: 仅管理员可传 `all`
      responses:
        '200':
          description: 邮件列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailListResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
...
  /api/forward-rules:
    get:
      tags: [ForwardRules]
      summary: 获取转发规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 规则列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardRuleListResponse'
        '403': { $ref: '#/components/responses/ForbiddenError' }
    post:
      tags: [ForwardRules]
      summary: 创建转发规则
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRuleInput'
      responses:
        '200': { $ref: '#/components/schemas/ForwardRuleResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/forward-rules/{id}:
    get:
      tags: [ForwardRules]
      summary: 规则详情
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      responses:
        '200': { $ref: '#/components/schemas/ForwardRuleResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    put:
      tags: [ForwardRules]
      summary: 更新规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRuleInput'
      responses:
        '200': { $ref: '#/components/schemas/BasicResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [ForwardRules]
      summary: 删除规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      responses:
        '200': { $ref: '#/components/schemas/BasicResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_cookies
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    ForwardRuleIdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    BasicResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: 123456 }
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string }
                user:
                  $ref: '#/components/schemas/UserProfile'
...
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        user_type: { type: integer, enum: [1], description: 固定为管理员 }
        created_at: { type: string }
        updated_at: { type: string }
        settings:
          type: object
          properties:
            webhook_url: { type: string, nullable: true }
            webhook_secret: { type: string, nullable: true }
...
    EmailListResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total: { type: integer }
                items:
                  type: array
                  items: { $ref: '#/components/schemas/EmailSummary' }
    EmailSummary:
      type: object
      properties:
        id: { type: string }
        subject: { type: string }
        sender_email: { type: string }
        recipient_email: { type: string }
        received_at: { type: string }
        has_attachments: { type: integer, enum: [0,1] }
...
    ForwardRuleInput:
      type: object
      required: [rule_name]
      properties:
        rule_name: { type: string }
        sender_filter: { type: string }
        keyword_filter: { type: string }
        recipient_filter: { type: string }
        enabled: { type: integer, enum: [0, 1] }
        webhooks:
          type: array
          description: 规则关联的多个 webhook 配置
          items:
            $ref: '#/components/schemas/ForwardRuleWebhook'
    ForwardRuleResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                rule: { $ref: '#/components/schemas/ForwardRule' }
    ForwardRuleListResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total: { type: integer }
                items:
                  type: array
                  items: { $ref: '#/components/schemas/ForwardRule' }
    ForwardRule:
      type: object
      properties:
        id: { type: integer }
        rule_name: { type: string }
        sender_filter: { type: string }
        keyword_filter: { type: string }
        recipient_filter: { type: string }
        webhook_url: { type: string, description: 保留字段以兼容旧数据 }
        webhook_secret: { type: string, description: 保留字段以兼容旧数据 }
        webhook_type: { type: string, enum: [dingtalk, feishu, bark, custom], description: 保留字段以兼容旧数据 }
        enabled: { type: integer, enum: [0, 1] }
        created_at: { type: string }
        updated_at: { type: string }
        webhooks:
          type: array
          description: 规则关联的多个 webhook 配置
          items:
            $ref: '#/components/schemas/ForwardRuleWebhook'
    ForwardRuleWebhook:
      type: object
      properties:
        id: { type: integer }
        rule_id: { type: integer }
        webhook_url: { type: string }
        webhook_secret: { type: string }
        webhook_type: { type: string, enum: [dingtalk, feishu, bark, custom] }
        enabled: { type: integer, enum: [0, 1] }
        created_at: { type: string }
        updated_at: { type: string }
...
  responses:
    UnauthorizedError:
      description: 未登录或会话失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    ForbiddenError:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    NotFoundError:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'